Move Graph and SubgraphBuilder into example/
