+++ src/example/concat.rs
@@ -1,13 +1,14 @@
 use std::rc::Rc;
 use std::cell::RefCell;
 
-use progress::{Timestamp, Graph, Scope};
+use progress::{Timestamp, Scope};
 use progress::nested::Source::ScopeOutput;
 use progress::nested::Target::ScopeInput;
 use progress::count_map::CountMap;
 
 use communication::*;
 use communication::channels::ObserverHelper;
+use example::builder::Graph;
 use example::stream::Stream;
 use columnar::Columnar;
