+++ src/example/distinct.rs
@@ -3,9 +3,9 @@ use std::collections::hash_state::DefaultState;
 use std::hash::{hash, Hash, SipHasher};
 use std::default::Default;
 
-use progress::Graph;
 use communication::*;
 use communication::pact::Exchange;
+use example::builder::Graph;
 use example::stream::Stream;
 use example::unary::UnaryExt;
