+++ src/example/graph_builder.rs
@@ -4,12 +4,12 @@ use std::rc::Rc;
 use std::cell::RefCell;
 use std::marker::PhantomData;
 
-use progress::{Timestamp, Graph, CountMap};
+use progress::{Timestamp, CountMap};
 use progress::nested::subgraph::Source::{GraphInput, ScopeOutput};
 use progress::nested::subgraph::Target::{GraphOutput, ScopeInput};
 use progress::nested::product::Product;
-use progress::nested::builder::Builder as SubgraphBuilder;
 
+use example::builder::{Graph, SubgraphBuilder};
 use example::stream::Stream;
 use communication::*;
 use communication::channels::ObserverHelper;
