+++ src/example_static/enterleave.rs
@@ -4,7 +4,7 @@ use std::rc::Rc;
 use std::cell::RefCell;
 use std::marker::PhantomData;
 
-use progress::{Timestamp, Graph, CountMap};
+use progress::{Timestamp, CountMap};
 use progress::nested::subgraph::Source::{GraphInput, ScopeOutput};
 use progress::nested::subgraph::Target::{GraphOutput, ScopeInput};
 use progress::nested::product::Product;
