+++ src/progress/nested/mod.rs
@@ -7,4 +7,3 @@ pub mod scope_wrapper;
 pub mod summary;
 pub mod product;
 pub mod subgraph;
-pub mod builder;
